refactor: metodo menorQue en Fraccion

+++ evaluaciones/retos/reto004/Fraccion.java
@@ -33,10 +33,10 @@ public class Fraccion {
 
 
     public boolean menorQue(Fraccion fraccion){
-        return (long) numerador * fraccion.denominador < (long) fraccion.numerador * denominador;
+        return numerador * fraccion.denominador < fraccion.numerador * denominador;
     }
     public boolean mayorQue(Fraccion fraccion){
-
+        
     }
